Split the peephole optimizer into an ordered pass pipeline
optimize.rs held three unrelated rewrites behind one
is_removable_pair -> bool predicate, feeding a driver that marks both
instructions of a matched pair as removed. Reshape it the way the
analyzer is arranged: a driver owning an ordered sequence of named
passes, each in its own module with its own parameters, over shared
rewrite machinery.

- optimize/rewrite.rs holds everything around a pass's decision: decode,
  jump-target protection, byte-stream rebuild, JMP/JMP_IF_NOT/CMP_BR
  offset rewriting and offset-map construction.
- optimize/pass_self_assign.rs and optimize/pass_arith_identity.rs each
  supply only their match on a two-instruction window. As with the
  analyzer's transforms, they take what they need and no more: the
  self-assignment pass takes no constant pool at all.
- optimize/mod.rs sequences the passes and composes their offset maps,
  which preserves the snap-forward property remap_line_map relies on.

Behaviour is unchanged. The existing tests move verbatim into
optimize/tests.rs with only their imports adjusted; they pin the
driver's observable output, which is exactly what must not move.
Compiling sample programs before and after produces byte-identical
containers.

Two tests are added for offset-map composition, the one mechanism that
is genuinely new because there was previously only ever one map.

The passes match on disjoint opcode pairs, so their order does not
affect the result today. The order is named and commented anyway, so a
future pass whose output feeds another has somewhere to say so.

Slice 1 of #1509.
